Refactor DownloadResponseTrait to be a new class, DownloadResponse
This takes advice from weierophinney that I should create a class called
DownloadResponse which builds on DownloadResponseTrait. This is the
first part of the refactor that changes enough of the functionality to
keep the code functional. I'm planning to implement the follow-up advice
as well, but one step at a time.