Add functionality to create CSV response from array and traversable
In addition to the refactor to create the CSV response from a string,
this change adds the ability to create the CSV response from a plain old
PHP array as well as a Traversable. I'm not altogether happy with the
naming of some of the functions and variables, but will address that,
just as soon as I can think of better names.